chore(tests): remove ApiClient tests to streamline test suite and reduce redundancy
